Moved default provider implementations to test folder to serve as examples. Their dependencies were also removed.
