+++ package.json
@@ -36,19 +36,6 @@
   "dependencies": {
     "atlassian-jwt": "^2.0.2"
   },
-  "peerDependencies": {
-    "axios": "*",
-    "got": "^11.8.2",
-    "node-fetch": "^2.6.5"
-  },
-  "peerDependenciesMeta": {
-    "axios": {
-      "optional": true
-    },
-    "node-fetch": {
-      "optional": true
-    }
-  },
   "devDependencies": {
     "@types/express": "^4.17.13",
     "@types/jest": "^27.0.2",
