+++ src/publicKeyProvider/index.ts
@@ -1,3 +1,2 @@
 export * from './CachingKeyProvider';
-export * from './GotKeyProvider';
 export * from './KeyProvider';
